cache line opt

+++ Seat.java
@@ -4,7 +4,7 @@ import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
 
 public class Seat {
-	volatile int seat_state;
+	volatile long seat_state;
 	final Lock seatlock = new ReentrantLock();
 	public int stationnum;
 	
@@ -14,19 +14,19 @@ public class Seat {
 	}
 	
 	//validate a seat
-	public boolean isEmpty(int ticket) {
+	public boolean isEmpty(long ticket) {
 		if((seat_state & ticket) == 0)
 			return true;
 		return false;
 	}
 	
 	//buy a ticket on this seat
-	public void buy(int ticket) {
+	public void buy(long ticket) {
 		seat_state = seat_state | ticket;
 	}
 	
 	//refund a ticket on this seat
-	public void refund(int ticket) {
+	public void refund(long ticket) {
 		seat_state = seat_state ^ ticket;
 		
 	}
